refactor: 에러 핸들링 로직 리팩토링
- fetchWithErrorHandling 함수명 역할에 맞게 내부에서 에러를 잡아 분기 처리함
- 외부에서는 try-catch로 에러 핸들링하는 것이 아닌 단순히 에러를 화면에 보여주는 UI 로직만 존재하도록 함
- 타입 가드를 통해 정상적인 경우 MoviesResponse, 에러인 경우 error를 포함한 객체가 있는지 확인
